// src/FE/src/api/generated/models/conversationActionStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type ConversationActionStatus = typeof ConversationActionStatus[keyof typeof ConversationActionStatus];


export const ConversationActionStatus = {
  Pending: 'Pending',
  Completed: 'Completed',
} as const;