// src/FE/src/api/generated/models/conversationMessageResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ConversationActionResponse } from './conversationActionResponse';

export interface ConversationMessageResponse {
  id: string;
  jobId: string;
  authorUserId: string;
  authorDisplayName: string;
  body: string;
  mentionedUserIds: string[];
  action: null | ConversationActionResponse;
  createdUtc: string;
}